% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PSDexperiment.R
\name{load_PSDExperiment}
\alias{load_PSDExperiment}
\title{Load PSD-IV instrument raw data}
\usage{
load_PSDExperiment(numeric_range, dets = NA, type)
}
\arguments{
\item{numeric_range, }{vector of integers representing the experiments to pull into the data space; for example c(1, 3, 10:12) will load data sets associated with 5 non-sequential experiments.}

\item{dets, }{vector of integers representing the detectors used; defaults to \emph{NA} to include all detectors; for example c(1:13, 15) would include all detectors numbered 1 through 15 except for number 14.}

\item{type, }{character string representing the return data type; for example "pos" would return a position spectrum.}
}
\value{
A list conaining recursive lists of data broken down on a per-detector, per-segment, per-scan, per-experiment basis.
}
\description{
Use this function to load raw data from the PSD-IV instrument. It is assumed that the data is in the current working directory and has been generated by an Instrumentation Associates implementation.
}
\details{
\emph{numeric_range} is a singular value or vector of values composed of integer numbers that represent the experiment number. The experiments are stored in a directory structure that is predefined by Instrumentation Associates in the form: ~\\ExptAAAAA\\ScanBB\\SegCC\\detDD.raw, where AAAAA is the experiment number, BB is the scan sequence number, CC is the angular segment, and DD is the detector number. The user specifies the desired experiments, and the function automatically traverses the directories and stores results in a complex list of lists.

\emph{dets} is a singular value or vector of values composed of integer numbers that represent the detector index numbers. Instrumentation Associates uses a base-zero numbering system, however the software uses a base-1 system for the convenience of the user. Suitable numbers are positive integers.

\emph{type} is a character string that determines the type of data to return. Possible values are "pos" for a position spectrum, "ab" for a- and b-channel pulse height histograms, or "sum" for the summation of the a- and b-side channels.
}
\author{
Peter Norgard
}
