% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PSDexperiment.R
\name{load_PSDExperiment}
\alias{load_PSDExperiment}
\title{Load PSD-IV experimental data.}
\usage{
load_PSDExperiment(numeric_range, dets = NA, type)
}
\arguments{
\item{numeric_range, }{vector of integers representing the experiments to pull into the data space}

\item{dets, }{vector of integers representing the detectors used; defaults to \emph{NA} to include all detectors}

\item{type, }{character string representing the return data type}
}
\description{
Use this function to load data from the PSD-IV instrument. It is assumed that
the data is in the current working directory and has been generated by an
Instrumentation Associates (IA) implementation. Data will be output on a per-
detector basis into a specific directory structure.
}
\author{
Peter Norgard
}
